Payments Endpoint (#11)
- Added handlers for GET /emoji_id/{yat}/payment API call
- Fixed issue related to percent encoding
